refactor(vep): remove lookup_variants/VariantLookupExec; Lance-only docs
Remove the legacy lookup_variants UDTF and its VariantLookupExec interval-join
(a second lookup path never used by annotate_vep, which always takes the Lance
KvLookupExec path): drop the UDTF registration, table_function.rs, the
ExecutionPlan + streaming machinery, the LookupProvider parquet branch and its
parquet-path tests, and the lookup_parquet_bench example. Move the live
colocated-cache-entry types and allele-comparison/genomic-shift helpers the
Lance path imports into a new colocated.rs (verbatim); importers updated.

Rewrite the README VEP section around the single Lance-backed annotate_vep
entrypoint and point the Ensembl VEP comparison at vepyr run_annotation_fast.py.

No downstream consumer references lookup_variants (verified: polars-bio,
comet-bio, vepyr). 748 lib tests pass; clippy --all-targets -D warnings clean.
