+++ datafusion/bio-function-vep/src/lookup_provider.rs
@@ -1,9 +1,7 @@
-//! Lookup provider for `lookup_variants()` table function.
+//! Lookup provider for the Lance variation cache.
 //!
-//! Implements variant lookup using a self-contained left interval join
-//! (`VariantLookupExec`) that builds VCF rows into per-chromosome COITrees
-//! and streams cache rows as the probe side. When the cache is a KV store,
-//! dispatches to `KvLookupExec` instead.
+//! Builds a `KvLookupExec` (Lance backend) over the per-chrom variation cache;
+//! consumed by the `annotate_vep` annotation path.
 
 use std::any::Any;
 use std::fmt::{Debug, Formatter};
@@ -14,14 +12,14 @@ use async_trait::async_trait;
 use datafusion::arrow::array::{Array, StringArray};
 use datafusion::arrow::datatypes::{DataType, Field, Schema, SchemaRef};
 use datafusion::catalog::Session;
-use datafusion::common::Result;
+use datafusion::common::{DataFusionError, Result};
 use datafusion::datasource::{TableProvider, TableType};
 use datafusion::physical_plan::ExecutionPlan;
 use datafusion::prelude::{Expr, SessionContext};
 
+use crate::colocated::ColocatedSink;
 use crate::coordinate::CoordinateNormalizer;
 use crate::schema_contract::validate_variation_schema;
-use crate::variant_lookup_exec::{ColocatedSink, VariantLookupExec};
 
 /// Wrap an execution plan with a `ProjectionExec` if projection is requested.
 ///
@@ -53,22 +51,13 @@ fn wrap_with_projection(
     }
 }
 
-/// Table provider that implements variant lookup via self-contained left
-/// interval join (`VariantLookupExec`).
+/// Table provider that implements variant lookup against the Lance variation
+/// cache via `KvLookupExec`.
 ///
-/// VCF variants are collected into per-chromosome COITrees (build side).
-/// Cache rows are streamed as the probe side with `match_allele()` as a
-/// post-filter. Unmatched VCF rows appear with NULL cache columns.
-///
-/// # Default mode (equi-join)
-///
-/// Requires VCF and cache coordinates to match exactly after VEP
-/// normalization (prefix/suffix stripping). SNVs and simple indels work well.
-///
-/// # Extended probes mode
-///
-/// When `extended_probes = true`, uses interval-overlap matching to handle
-/// VEP-style coordinate encodings (insertion-style start > end,
+/// VCF variants are streamed and probed against the per-chrom Lance variation
+/// dataset; `match_allele()` is applied as a post-filter and unmatched VCF rows
+/// appear with NULL cache columns. `extended_probes = true` uses interval-overlap
+/// matching for VEP-style coordinate encodings (insertion-style start > end,
 /// prefix-trimmed deletion shifts, tandem-repeat right-shifting).
 pub struct LookupProvider {
     session: Arc<SessionContext>,
@@ -91,7 +80,7 @@ pub struct LookupProvider {
     schema: SchemaRef,
     /// Optional sink for co-located data collection during probe phase.
     colocated_sink: Option<ColocatedSink>,
-    /// Optional partition-local sinks for fjall co-located data collection.
+    /// Optional partition-local sinks for co-located data collection.
     partition_colocated_sinks: Option<Vec<ColocatedSink>>,
     /// Root of a Lance variation cache. When set, variation lookup uses
     /// `variation.lance/chrN.lance` plus the sidecar position/bloom indexes.
@@ -179,7 +168,7 @@ impl LookupProvider {
         self.colocated_sink = Some(sink);
     }
 
-    /// Set partition-local co-located data sinks for fjall lookup execution.
+    /// Set partition-local co-located data sinks for lance lookup execution.
     pub fn set_partition_colocated_sinks(&mut self, sinks: Vec<ColocatedSink>) {
         self.partition_colocated_sinks = Some(sinks);
     }
@@ -294,933 +283,10 @@ impl TableProvider for LookupProvider {
             return wrap_with_projection(plan, projection);
         }
 
-        // Parquet / MemTable path: use VariantLookupExec.
-        let vcf_has_chr = has_chr_prefix(&self.session, &self.vcf_table).await?;
-
-        let vcf_df = self.session.table(&self.vcf_table).await?;
-        let vcf_df = if let Some(ref filter) = self.vcf_filter {
-            vcf_df.filter(filter.clone())?
-        } else {
-            vcf_df
-        };
-        let vcf_plan = vcf_df.create_physical_plan().await?;
-
-        // Project cache to only the columns needed: join keys + requested output columns.
-        // This avoids reading all 78 parquet columns when only ~10 are needed.
-        let cache_df = self.session.table(&self.cache_table).await?;
-        let mut required_cols: Vec<&str> = vec!["chrom", "start", "end", "allele_string"];
-        // Only request "failed" if the cache schema actually contains it.
-        if cache_df
-            .schema()
-            .as_arrow()
-            .field_with_name("failed")
-            .is_ok()
-        {
-            required_cols.push("failed");
-        }
-        for col in &self.cache_columns {
-            if !required_cols.contains(&col.as_str()) {
-                required_cols.push(col.as_str());
-            }
-        }
-        let cache_plan = cache_df
-            .select_columns(&required_cols)?
-            .create_physical_plan()
-            .await?;
-
-        let mut exec = VariantLookupExec::new(
-            vcf_plan,
-            cache_plan,
-            self.cache_columns.clone(),
-            vcf_has_chr,
-            self.coord_normalizer.clone(),
-            self.extended_probes,
-            self.allowed_failed,
-            self.reference_fasta_path.clone(),
-            self.schema.clone(),
-        );
-        if let Some(ref sink) = self.colocated_sink {
-            exec = exec.with_colocated_sink(Arc::clone(sink));
-        }
-        let plan: Arc<dyn ExecutionPlan> = Arc::new(exec);
-        wrap_with_projection(plan, projection)
+        // Lance is the only supported variation-lookup backend; a LookupProvider
+        // is always constructed with a Lance cache root by the annotation path.
+        Err(DataFusionError::Plan(
+            "LookupProvider::scan requires a Lance variation cache root".to_string(),
+        ))
     }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::LookupProvider;
-    use crate::create_vep_session;
-    use datafusion::arrow::array::{
-        Array, ArrayRef, Int64Array, RecordBatch, StringArray, StringViewArray,
-    };
-    use datafusion::arrow::datatypes::{DataType, Field, Schema};
-    use datafusion::datasource::MemTable;
-    use datafusion::physical_plan::displayable;
-    use std::collections::HashMap;
-    use std::sync::Arc;
-
-    /// Extract string values from an array that may be Utf8, Utf8View, or LargeUtf8.
-    fn string_values(col: &ArrayRef) -> Vec<Option<String>> {
-        use datafusion::arrow::array::LargeStringArray;
-        if let Some(arr) = col.as_any().downcast_ref::<StringArray>() {
-            (0..arr.len())
-                .map(|i| {
-                    if arr.is_null(i) {
-                        None
-                    } else {
-                        Some(arr.value(i).to_string())
-                    }
-                })
-                .collect()
-        } else if let Some(arr) = col.as_any().downcast_ref::<StringViewArray>() {
-            (0..arr.len())
-                .map(|i| {
-                    if arr.is_null(i) {
-                        None
-                    } else {
-                        Some(arr.value(i).to_string())
-                    }
-                })
-                .collect()
-        } else if let Some(arr) = col.as_any().downcast_ref::<LargeStringArray>() {
-            (0..arr.len())
-                .map(|i| {
-                    if arr.is_null(i) {
-                        None
-                    } else {
-                        Some(arr.value(i).to_string())
-                    }
-                })
-                .collect()
-        } else {
-            panic!(
-                "expected Utf8, Utf8View, or LargeUtf8, got {:?}",
-                col.data_type()
-            );
-        }
-    }
-
-    fn schema_with_coord_metadata(fields: Vec<Field>, zero_based: bool) -> Arc<Schema> {
-        let mut metadata = HashMap::new();
-        metadata.insert(
-            "bio.coordinate_system_zero_based".to_string(),
-            zero_based.to_string(),
-        );
-        Arc::new(Schema::new_with_metadata(fields, metadata))
-    }
-
-    fn vcf_table() -> MemTable {
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("ref", DataType::Utf8, false),
-            Field::new("alt", DataType::Utf8, false),
-        ]));
-        let batch = RecordBatch::try_new(
-            schema.clone(),
-            vec![
-                // Three VCF rows: first two match cache, third has no match
-                Arc::new(StringArray::from(vec!["chr1", "chr1", "chr2"])),
-                Arc::new(Int64Array::from(vec![100, 200, 500])),
-                Arc::new(Int64Array::from(vec![101, 201, 501])),
-                Arc::new(StringArray::from(vec!["A", "C", "G"])),
-                Arc::new(StringArray::from(vec!["G", "T", "A"])),
-            ],
-        )
-        .unwrap();
-        MemTable::try_new(schema, vec![vec![batch]]).unwrap()
-    }
-
-    fn cache_table() -> MemTable {
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("variation_name", DataType::Utf8, true),
-            Field::new("allele_string", DataType::Utf8, false),
-            Field::new("clin_sig", DataType::Utf8, true),
-            Field::new("failed", DataType::Int64, false),
-        ]));
-        // Cache coordinates must exactly match VCF after normalization.
-        // VCF has chr1:(100,101) and chr1:(200,201), cache uses bare chrom names.
-        let batch = RecordBatch::try_new(
-            schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1", "1"])),
-                Arc::new(Int64Array::from(vec![100, 200])),
-                Arc::new(Int64Array::from(vec![101, 201])),
-                Arc::new(StringArray::from(vec!["rs123", "rs456"])),
-                Arc::new(StringArray::from(vec!["A/G", "C/T"])),
-                Arc::new(StringArray::from(vec!["benign", "pathogenic"])),
-                Arc::new(Int64Array::from(vec![0, 0])),
-            ],
-        )
-        .unwrap();
-        MemTable::try_new(schema, vec![vec![batch]]).unwrap()
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_generates_variant_lookup_exec() {
-        let ctx = create_vep_session();
-
-        ctx.register_table("vcf_data", Arc::new(vcf_table()))
-            .unwrap();
-        ctx.register_table("var_cache", Arc::new(cache_table()))
-            .unwrap();
-
-        let df = ctx
-            .sql("SELECT * FROM lookup_variants('vcf_data', 'var_cache')")
-            .await
-            .unwrap();
-
-        let plan = df.create_physical_plan().await.unwrap();
-        let plan_str = displayable(plan.as_ref()).indent(true).to_string();
-
-        assert!(
-            plan_str.contains("VariantLookupExec"),
-            "Expected VariantLookupExec in plan, got:\n{plan_str}"
-        );
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_left_join_preserves_unmatched_vcf_rows() {
-        let ctx = create_vep_session();
-
-        ctx.register_table("vcf_data", Arc::new(vcf_table()))
-            .unwrap();
-        ctx.register_table("var_cache", Arc::new(cache_table()))
-            .unwrap();
-
-        let df = ctx
-            .sql("SELECT * FROM lookup_variants('vcf_data', 'var_cache', 'clin_sig')")
-            .await
-            .unwrap();
-
-        // Verify the plan uses VariantLookupExec
-        let plan = df.clone().create_physical_plan().await.unwrap();
-        let plan_str = displayable(plan.as_ref()).indent(true).to_string();
-        assert!(
-            plan_str.contains("VariantLookupExec"),
-            "Expected VariantLookupExec in plan, got:\n{plan_str}"
-        );
-
-        let batches = df.collect().await.unwrap();
-
-        // Count total rows — should be 3 (two matches + one unmatched VCF row)
-        let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-        assert_eq!(
-            total_rows, 3,
-            "Expected 3 rows (2 matched + 1 unmatched), got {total_rows}"
-        );
-
-        // Collect all chrom values and cache_clin_sig values
-        let mut chroms = Vec::new();
-        let mut clin_sigs = Vec::new();
-        for batch in &batches {
-            let chrom_col = batch.column_by_name("chrom").unwrap();
-            let clin_sig_col = batch.column_by_name("cache_clin_sig").unwrap();
-            let chrom_vals = string_values(chrom_col);
-            let clin_sig_vals = string_values(clin_sig_col);
-            for i in 0..batch.num_rows() {
-                chroms.push(chrom_vals[i].clone().unwrap_or_else(|| "NULL".to_string()));
-                clin_sigs.push(
-                    clin_sig_vals[i]
-                        .clone()
-                        .unwrap_or_else(|| "NULL".to_string()),
-                );
-            }
-        }
-
-        // The chr2 variant has no cache match — it should appear with NULL annotation.
-        // Output keeps the original VCF chrom value ("chr2").
-        assert!(
-            chroms.contains(&"chr2".to_string()),
-            "Expected 'chr2' (unmatched VCF row) in output, got chroms: {chroms:?}"
-        );
-
-        let chr2_idx = chroms.iter().position(|c| c == "chr2").unwrap();
-        assert_eq!(
-            clin_sigs[chr2_idx], "NULL",
-            "Expected NULL clin_sig for unmatched chr2 row, got: {}",
-            clin_sigs[chr2_idx]
-        );
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_preserves_vcf_row_order_when_cache_rows_are_reversed() {
-        let ctx = create_vep_session();
-
-        let vcf_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("ref", DataType::Utf8, false),
-            Field::new("alt", DataType::Utf8, false),
-        ]));
-        let vcf_batch = RecordBatch::try_new(
-            vcf_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1", "1", "1"])),
-                Arc::new(Int64Array::from(vec![100, 200, 300])),
-                Arc::new(Int64Array::from(vec![100, 200, 300])),
-                Arc::new(StringArray::from(vec!["A", "C", "G"])),
-                Arc::new(StringArray::from(vec!["T", "A", "C"])),
-            ],
-        )
-        .unwrap();
-        let vcf = MemTable::try_new(vcf_schema, vec![vec![vcf_batch]]).unwrap();
-
-        let cache_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("allele_string", DataType::Utf8, false),
-            Field::new("variation_name", DataType::Utf8, true),
-            Field::new("failed", DataType::Int64, false),
-        ]));
-        let cache_batch = RecordBatch::try_new(
-            cache_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1", "1", "1"])),
-                Arc::new(Int64Array::from(vec![300, 200, 100])),
-                Arc::new(Int64Array::from(vec![300, 200, 100])),
-                Arc::new(StringArray::from(vec!["G/C", "C/A", "A/T"])),
-                Arc::new(StringArray::from(vec!["rs300", "rs200", "rs100"])),
-                Arc::new(Int64Array::from(vec![0, 0, 0])),
-            ],
-        )
-        .unwrap();
-        let cache = MemTable::try_new(cache_schema, vec![vec![cache_batch]]).unwrap();
-
-        ctx.register_table("vcf_ordered", Arc::new(vcf)).unwrap();
-        ctx.register_table("cache_reversed", Arc::new(cache))
-            .unwrap();
-
-        let batches = ctx
-            .sql("SELECT * FROM lookup_variants('vcf_ordered', 'cache_reversed', 'variation_name')")
-            .await
-            .unwrap()
-            .collect()
-            .await
-            .unwrap();
-
-        let mut starts = Vec::new();
-        let mut variation_names = Vec::new();
-        for batch in &batches {
-            let start_col = batch
-                .column_by_name("start")
-                .unwrap()
-                .as_any()
-                .downcast_ref::<Int64Array>()
-                .unwrap();
-            let names = string_values(batch.column_by_name("cache_variation_name").unwrap());
-            for row in 0..batch.num_rows() {
-                starts.push(start_col.value(row));
-                variation_names.push(names[row].clone().unwrap());
-            }
-        }
-
-        assert_eq!(starts, vec![100, 200, 300]);
-        assert_eq!(variation_names, vec!["rs100", "rs200", "rs300"]);
-    }
-
-    /// Regression test: equi-join ensures VCF rows only match cache entries
-    /// at exactly the same (chrom, start, end) coordinates. Decoy entries at
-    /// adjacent positions must NOT produce output rows.
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_no_false_matches_at_adjacent_positions() {
-        let ctx = create_vep_session();
-
-        // VCF: 1-based closed SNVs (start=end for point variants)
-        let vcf_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("ref", DataType::Utf8, false),
-            Field::new("alt", DataType::Utf8, false),
-        ]));
-        let vcf_batch = RecordBatch::try_new(
-            vcf_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1", "1"])),
-                Arc::new(Int64Array::from(vec![100, 200])),
-                Arc::new(Int64Array::from(vec![100, 200])),
-                Arc::new(StringArray::from(vec!["A", "C"])),
-                Arc::new(StringArray::from(vec!["G", "T"])),
-            ],
-        )
-        .unwrap();
-        let vcf = MemTable::try_new(vcf_schema, vec![vec![vcf_batch]]).unwrap();
-
-        // VEP cache: 1-based closed intervals (start=end for SNVs).
-        // Includes TRUE matches at positions 100 and 200, PLUS decoy entries
-        // at adjacent positions (101, 201) with the same ALT allele but
-        // different REF — these must NOT produce output rows.
-        let cache_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("variation_name", DataType::Utf8, true),
-            Field::new("allele_string", DataType::Utf8, false),
-            Field::new("clin_sig", DataType::Utf8, true),
-            Field::new("failed", DataType::Int64, false),
-        ]));
-        let cache_batch = RecordBatch::try_new(
-            cache_schema.clone(),
-            vec![
-                // chrom:  100 match, 101 decoy, 200 match, 201 decoy
-                Arc::new(StringArray::from(vec!["1", "1", "1", "1"])),
-                Arc::new(Int64Array::from(vec![100, 101, 200, 201])),
-                Arc::new(Int64Array::from(vec![100, 101, 200, 201])),
-                Arc::new(StringArray::from(vec!["rs100", "rs101", "rs200", "rs201"])),
-                // True match alleles + decoy alleles (same ALT, different REF)
-                Arc::new(StringArray::from(vec!["A/G", "T/G", "C/T", "G/T"])),
-                Arc::new(StringArray::from(vec![
-                    "benign",
-                    "decoy_101",
-                    "pathogenic",
-                    "decoy_201",
-                ])),
-                Arc::new(Int64Array::from(vec![0, 0, 0, 0])),
-            ],
-        )
-        .unwrap();
-        let cache = MemTable::try_new(cache_schema, vec![vec![cache_batch]]).unwrap();
-
-        ctx.register_table("vcf_narrow", Arc::new(vcf)).unwrap();
-        ctx.register_table("cache_narrow", Arc::new(cache)).unwrap();
-
-        let df = ctx
-            .sql("SELECT * FROM lookup_variants('vcf_narrow', 'cache_narrow', 'variation_name,clin_sig')")
-            .await
-            .unwrap();
-        let batches = df.collect().await.unwrap();
-
-        let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-        assert_eq!(
-            total_rows, 2,
-            "Expected exactly 2 rows (one match per VCF variant, no decoy matches), got {total_rows}"
-        );
-
-        // Verify only the true matches appear (not the decoys)
-        let mut var_names = Vec::new();
-        for batch in &batches {
-            let col = batch.column_by_name("cache_variation_name").unwrap();
-            for val in string_values(col) {
-                if let Some(v) = val {
-                    var_names.push(v);
-                }
-            }
-        }
-        assert!(
-            var_names.contains(&"rs100".to_string()),
-            "Expected rs100 in output, got: {var_names:?}"
-        );
-        assert!(
-            var_names.contains(&"rs200".to_string()),
-            "Expected rs200 in output, got: {var_names:?}"
-        );
-        assert!(
-            !var_names.contains(&"rs101".to_string()),
-            "Decoy rs101 should NOT appear in output, got: {var_names:?}"
-        );
-        assert!(
-            !var_names.contains(&"rs201".to_string()),
-            "Decoy rs201 should NOT appear in output, got: {var_names:?}"
-        );
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_matches_pipe_joined_multi_alt_input() {
-        let ctx = create_vep_session();
-
-        let vcf_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("ref", DataType::Utf8, false),
-            Field::new("alt", DataType::Utf8, false),
-        ]));
-        let vcf_batch = RecordBatch::try_new(
-            vcf_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1"])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(StringArray::from(vec!["A"])),
-                Arc::new(StringArray::from(vec!["G|T"])),
-            ],
-        )
-        .unwrap();
-        let vcf = MemTable::try_new(vcf_schema, vec![vec![vcf_batch]]).unwrap();
-
-        let cache_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("variation_name", DataType::Utf8, true),
-            Field::new("allele_string", DataType::Utf8, false),
-            Field::new("clin_sig", DataType::Utf8, true),
-            Field::new("failed", DataType::Int64, false),
-        ]));
-        let cache_batch = RecordBatch::try_new(
-            cache_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1", "1"])),
-                Arc::new(Int64Array::from(vec![100, 100])),
-                Arc::new(Int64Array::from(vec![100, 100])),
-                Arc::new(StringArray::from(vec!["rs_g_match", "rs_c_miss"])),
-                Arc::new(StringArray::from(vec!["A/G", "A/C"])),
-                Arc::new(StringArray::from(vec!["benign", "benign"])),
-                Arc::new(Int64Array::from(vec![0, 0])),
-            ],
-        )
-        .unwrap();
-        let cache = MemTable::try_new(cache_schema, vec![vec![cache_batch]]).unwrap();
-
-        ctx.register_table("vcf_multi_alt_pipe", Arc::new(vcf))
-            .unwrap();
-        ctx.register_table("cache_multi_alt_pipe", Arc::new(cache))
-            .unwrap();
-
-        let df = ctx
-            .sql("SELECT * FROM lookup_variants('vcf_multi_alt_pipe', 'cache_multi_alt_pipe', 'variation_name')")
-            .await
-            .unwrap();
-        let batches = df.collect().await.unwrap();
-
-        let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-        assert_eq!(total_rows, 1, "Expected one row from ALT=G|T matching A/G");
-
-        let mut var_names = Vec::new();
-        for batch in &batches {
-            let col = batch.column_by_name("cache_variation_name").unwrap();
-            for val in string_values(col) {
-                if let Some(v) = val {
-                    var_names.push(v);
-                }
-            }
-        }
-        assert_eq!(var_names, vec!["rs_g_match"]);
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_matches_insertion_style_cache_coordinates() {
-        let ctx = create_vep_session();
-
-        // One-based closed insertion at position 100: REF=A ALT=AT.
-        let vcf_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("ref", DataType::Utf8, false),
-            Field::new("alt", DataType::Utf8, false),
-        ]));
-        let vcf_batch = RecordBatch::try_new(
-            vcf_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1"])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(StringArray::from(vec!["A"])),
-                Arc::new(StringArray::from(vec!["AT"])),
-            ],
-        )
-        .unwrap();
-        let vcf = MemTable::try_new(vcf_schema, vec![vec![vcf_batch]]).unwrap();
-
-        // Cache uses insertion-style coordinates: start=end+1.
-        let cache_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("variation_name", DataType::Utf8, true),
-            Field::new("allele_string", DataType::Utf8, false),
-            Field::new("clin_sig", DataType::Utf8, true),
-            Field::new("failed", DataType::Int64, false),
-        ]));
-        let cache_batch = RecordBatch::try_new(
-            cache_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1", "1"])),
-                Arc::new(Int64Array::from(vec![101, 102])),
-                Arc::new(Int64Array::from(vec![100, 101])),
-                Arc::new(StringArray::from(vec!["rs_ins_match", "rs_ins_miss"])),
-                Arc::new(StringArray::from(vec!["-/T", "-/T"])),
-                Arc::new(StringArray::from(vec!["pathogenic", "pathogenic"])),
-                Arc::new(Int64Array::from(vec![0, 0])),
-            ],
-        )
-        .unwrap();
-        let cache = MemTable::try_new(cache_schema, vec![vec![cache_batch]]).unwrap();
-
-        ctx.register_table("vcf_insertion", Arc::new(vcf)).unwrap();
-        ctx.register_table("cache_insertion_style", Arc::new(cache))
-            .unwrap();
-
-        let df = ctx
-            .sql("SELECT * FROM lookup_variants('vcf_insertion', 'cache_insertion_style', 'variation_name', 'exact', true)")
-            .await
-            .unwrap();
-        let batches = df.collect().await.unwrap();
-
-        let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-        assert_eq!(
-            total_rows, 1,
-            "Expected one match against insertion-style cache coordinates"
-        );
-
-        let mut var_names = Vec::new();
-        for batch in &batches {
-            let col = batch.column_by_name("cache_variation_name").unwrap();
-            for val in string_values(col) {
-                if let Some(v) = val {
-                    var_names.push(v);
-                }
-            }
-        }
-        assert_eq!(var_names, vec!["rs_ins_match"]);
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_one_based_metadata_matches_point_variants() {
-        let ctx = create_vep_session();
-
-        // Explicitly mark both tables as 1-based closed.
-        let vcf_schema = schema_with_coord_metadata(
-            vec![
-                Field::new("chrom", DataType::Utf8, false),
-                Field::new("start", DataType::Int64, false),
-                Field::new("end", DataType::Int64, false),
-                Field::new("ref", DataType::Utf8, false),
-                Field::new("alt", DataType::Utf8, false),
-            ],
-            false,
-        );
-        let vcf_batch = RecordBatch::try_new(
-            vcf_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1"])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(StringArray::from(vec!["A"])),
-                Arc::new(StringArray::from(vec!["G"])),
-            ],
-        )
-        .unwrap();
-        let vcf = MemTable::try_new(vcf_schema, vec![vec![vcf_batch]]).unwrap();
-
-        let cache_schema = schema_with_coord_metadata(
-            vec![
-                Field::new("chrom", DataType::Utf8, false),
-                Field::new("start", DataType::Int64, false),
-                Field::new("end", DataType::Int64, false),
-                Field::new("variation_name", DataType::Utf8, true),
-                Field::new("allele_string", DataType::Utf8, false),
-                Field::new("clin_sig", DataType::Utf8, true),
-                Field::new("failed", DataType::Int64, false),
-            ],
-            false,
-        );
-        let cache_batch = RecordBatch::try_new(
-            cache_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1"])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(StringArray::from(vec!["rs100"])),
-                Arc::new(StringArray::from(vec!["A/G"])),
-                Arc::new(StringArray::from(vec!["benign"])),
-                Arc::new(Int64Array::from(vec![0])),
-            ],
-        )
-        .unwrap();
-        let cache = MemTable::try_new(cache_schema, vec![vec![cache_batch]]).unwrap();
-
-        ctx.register_table("vcf_one_based", Arc::new(vcf)).unwrap();
-        ctx.register_table("cache_one_based", Arc::new(cache))
-            .unwrap();
-
-        let df = ctx
-            .sql("SELECT * FROM lookup_variants('vcf_one_based', 'cache_one_based', 'variation_name')")
-            .await
-            .unwrap();
-        let batches = df.collect().await.unwrap();
-
-        let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-        assert_eq!(total_rows, 1, "Expected a single joined row");
-
-        let mut var_names = Vec::new();
-        for batch in &batches {
-            let col = batch.column_by_name("cache_variation_name").unwrap();
-            for val in string_values(col) {
-                if let Some(v) = val {
-                    var_names.push(v);
-                }
-            }
-        }
-        assert_eq!(var_names, vec!["rs100"]);
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_zero_based_input_to_one_based_cache_uses_metadata() {
-        let ctx = create_vep_session();
-
-        // VCF is 0-based half-open: position 100 is [99, 100).
-        let vcf_schema = schema_with_coord_metadata(
-            vec![
-                Field::new("chrom", DataType::Utf8, false),
-                Field::new("start", DataType::Int64, false),
-                Field::new("end", DataType::Int64, false),
-                Field::new("ref", DataType::Utf8, false),
-                Field::new("alt", DataType::Utf8, false),
-            ],
-            true,
-        );
-        let vcf_batch = RecordBatch::try_new(
-            vcf_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1"])),
-                Arc::new(Int64Array::from(vec![99])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(StringArray::from(vec!["A"])),
-                Arc::new(StringArray::from(vec!["G"])),
-            ],
-        )
-        .unwrap();
-        let vcf = MemTable::try_new(vcf_schema, vec![vec![vcf_batch]]).unwrap();
-
-        // Cache is 1-based closed. Both rows share allele_string, so the
-        // interval condition must discriminate the true position.
-        let cache_schema = schema_with_coord_metadata(
-            vec![
-                Field::new("chrom", DataType::Utf8, false),
-                Field::new("start", DataType::Int64, false),
-                Field::new("end", DataType::Int64, false),
-                Field::new("variation_name", DataType::Utf8, true),
-                Field::new("allele_string", DataType::Utf8, false),
-                Field::new("clin_sig", DataType::Utf8, true),
-                Field::new("failed", DataType::Int64, false),
-            ],
-            false,
-        );
-        let cache_batch = RecordBatch::try_new(
-            cache_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1", "1"])),
-                Arc::new(Int64Array::from(vec![100, 101])),
-                Arc::new(Int64Array::from(vec![100, 101])),
-                Arc::new(StringArray::from(vec!["rs100", "rs101"])),
-                Arc::new(StringArray::from(vec!["A/G", "A/G"])),
-                Arc::new(StringArray::from(vec!["benign", "benign"])),
-                Arc::new(Int64Array::from(vec![0, 0])),
-            ],
-        )
-        .unwrap();
-        let cache = MemTable::try_new(cache_schema, vec![vec![cache_batch]]).unwrap();
-
-        ctx.register_table("vcf_zero_based", Arc::new(vcf)).unwrap();
-        ctx.register_table("cache_closed", Arc::new(cache)).unwrap();
-
-        let df = ctx
-            .sql(
-                "SELECT * FROM lookup_variants('vcf_zero_based', 'cache_closed', 'variation_name')",
-            )
-            .await
-            .unwrap();
-        let batches = df.collect().await.unwrap();
-
-        let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-        assert_eq!(
-            total_rows, 1,
-            "Expected exactly one match after 0-based->1-based normalization"
-        );
-
-        let mut var_names = Vec::new();
-        for batch in &batches {
-            let col = batch.column_by_name("cache_variation_name").unwrap();
-            for val in string_values(col) {
-                if let Some(v) = val {
-                    var_names.push(v);
-                }
-            }
-        }
-        assert_eq!(var_names, vec!["rs100"]);
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_one_based_input_to_zero_based_cache_uses_metadata() {
-        let ctx = create_vep_session();
-
-        // VCF is 1-based closed: position 100 is [100, 100].
-        let vcf_schema = schema_with_coord_metadata(
-            vec![
-                Field::new("chrom", DataType::Utf8, false),
-                Field::new("start", DataType::Int64, false),
-                Field::new("end", DataType::Int64, false),
-                Field::new("ref", DataType::Utf8, false),
-                Field::new("alt", DataType::Utf8, false),
-            ],
-            false,
-        );
-        let vcf_batch = RecordBatch::try_new(
-            vcf_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1"])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(StringArray::from(vec!["A"])),
-                Arc::new(StringArray::from(vec!["G"])),
-            ],
-        )
-        .unwrap();
-        let vcf = MemTable::try_new(vcf_schema, vec![vec![vcf_batch]]).unwrap();
-
-        // Cache is 0-based half-open:
-        // pos 100 -> [99, 100), pos 101 -> [100, 101).
-        let cache_schema = schema_with_coord_metadata(
-            vec![
-                Field::new("chrom", DataType::Utf8, false),
-                Field::new("start", DataType::Int64, false),
-                Field::new("end", DataType::Int64, false),
-                Field::new("variation_name", DataType::Utf8, true),
-                Field::new("allele_string", DataType::Utf8, false),
-                Field::new("clin_sig", DataType::Utf8, true),
-                Field::new("failed", DataType::Int64, false),
-            ],
-            true,
-        );
-        let cache_batch = RecordBatch::try_new(
-            cache_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1", "1"])),
-                Arc::new(Int64Array::from(vec![99, 100])),
-                Arc::new(Int64Array::from(vec![100, 101])),
-                Arc::new(StringArray::from(vec!["rs100", "rs101"])),
-                Arc::new(StringArray::from(vec!["A/G", "A/G"])),
-                Arc::new(StringArray::from(vec!["benign", "benign"])),
-                Arc::new(Int64Array::from(vec![0, 0])),
-            ],
-        )
-        .unwrap();
-        let cache = MemTable::try_new(cache_schema, vec![vec![cache_batch]]).unwrap();
-
-        ctx.register_table("vcf_closed", Arc::new(vcf)).unwrap();
-        ctx.register_table("cache_zero_based", Arc::new(cache))
-            .unwrap();
-
-        let df = ctx
-            .sql(
-                "SELECT * FROM lookup_variants('vcf_closed', 'cache_zero_based', 'variation_name')",
-            )
-            .await
-            .unwrap();
-        let batches = df.collect().await.unwrap();
-
-        let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-        assert_eq!(
-            total_rows, 1,
-            "Expected exactly one match after 1-based->0-based normalization"
-        );
-
-        let mut var_names = Vec::new();
-        for batch in &batches {
-            let col = batch.column_by_name("cache_variation_name").unwrap();
-            for val in string_values(col) {
-                if let Some(v) = val {
-                    var_names.push(v);
-                }
-            }
-        }
-        assert_eq!(var_names, vec!["rs100"]);
-    }
-
-    /// Duplicate cache rows on (chrom, start, end, allele_string) with different
-    /// variation_names each produce a separate output row.
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_lookup_cache_duplicates_produce_separate_rows() {
-        let ctx = create_vep_session();
-
-        let vcf_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("ref", DataType::Utf8, false),
-            Field::new("alt", DataType::Utf8, false),
-        ]));
-        let vcf_batch = RecordBatch::try_new(
-            vcf_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1"])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(Int64Array::from(vec![100])),
-                Arc::new(StringArray::from(vec!["A"])),
-                Arc::new(StringArray::from(vec!["G"])),
-            ],
-        )
-        .unwrap();
-        let vcf = MemTable::try_new(vcf_schema, vec![vec![vcf_batch]]).unwrap();
-
-        // Cache with THREE duplicate rows at the same position + allele.
-        // Only variation_name differs (e.g. dbSNP + COSMIC at same position).
-        let cache_schema = Arc::new(Schema::new(vec![
-            Field::new("chrom", DataType::Utf8, false),
-            Field::new("start", DataType::Int64, false),
-            Field::new("end", DataType::Int64, false),
-            Field::new("variation_name", DataType::Utf8, true),
-            Field::new("allele_string", DataType::Utf8, false),
-            Field::new("clin_sig", DataType::Utf8, true),
-            Field::new("failed", DataType::Int64, false),
-        ]));
-        let cache_batch = RecordBatch::try_new(
-            cache_schema.clone(),
-            vec![
-                Arc::new(StringArray::from(vec!["1", "1", "1"])),
-                Arc::new(Int64Array::from(vec![100, 100, 100])),
-                Arc::new(Int64Array::from(vec![100, 100, 100])),
-                Arc::new(StringArray::from(vec!["rs100", "COSM123", "rs100_dup"])),
-                Arc::new(StringArray::from(vec!["A/G", "A/G", "A/G"])),
-                Arc::new(StringArray::from(vec!["benign", "benign", "benign"])),
-                Arc::new(Int64Array::from(vec![0, 0, 0])),
-            ],
-        )
-        .unwrap();
-        let cache = MemTable::try_new(cache_schema, vec![vec![cache_batch]]).unwrap();
-
-        ctx.register_table("vcf_dedup", Arc::new(vcf)).unwrap();
-        ctx.register_table("cache_dedup", Arc::new(cache)).unwrap();
-
-        let df = ctx
-            .sql("SELECT * FROM lookup_variants('vcf_dedup', 'cache_dedup', 'variation_name,clin_sig')")
-            .await
-            .unwrap();
-        let batches = df.collect().await.unwrap();
-
-        // Deduplicated: one output row per VCF variant even when multiple
-        // cache alleles match (the first match wins; duplicates produce
-        // identical annotation output).
-        let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-        assert_eq!(
-            total_rows, 1,
-            "Expected 1 row (deduplicated per VCF variant), got {total_rows}"
-        );
-
-        // Verify the first matched allele's variation_name is emitted.
-        let mut var_names = Vec::new();
-        for batch in &batches {
-            let col = batch.column_by_name("cache_variation_name").unwrap();
-            for val in string_values(col) {
-                if let Some(v) = val {
-                    var_names.push(v);
-                }
-            }
-        }
-        assert_eq!(var_names.len(), 1, "Expected exactly one variation_name");
-    }
-
-    // -----------------------------------------------------------------------
-    // KV (fjall) parity tests — mirrors of the Parquet-path tests above
-    // -----------------------------------------------------------------------
 }
